% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_adj.R
\name{corr_adj}
\alias{corr_adj}
\title{corr_adj: formats p value files created with corr matrices}
\usage{
corr_adj(corData, pData, outputCSV = F, filename = "what is wrong")
}
\arguments{
\item{corData}{dataframe of correlation coefficients in the form of a matrix}

\item{pData}{dataframe of p values associated with corData r values}

\item{outputCSV}{Should R output csv files. Default is False}

\item{filename}{string containing name of output file}
}
\value{
csv file with correlation and p values concatenated
}
\description{
corr_adj: formats p value files created with corr matrices
}
