% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_mat.R
\name{corr_mat}
\alias{corr_mat}
\title{Function used to compute correlation matrix}
\usage{
corr_mat(
  dat = dat,
  raw = F,
  rank = F,
  log = F,
  outputCSV = F,
  file1 = "correlation matrix file name",
  file2 = "p value file name"
)
}
\arguments{
\item{dat}{a data frame used to test correlations}

\item{raw}{true/false binary indicating type of transformation}

\item{rank}{true/false binary indicating type of transformation}

\item{log}{true/false binary indicating type of transformation}

\item{outputCSV}{Should R output csv files. Default is False}

\item{file1}{output file name for r values}

\item{file2}{output file name for p values}
}
\value{
returns two csv files one with the r values and one with p values
}
\description{
Function used to compute correlation matrix
}
\details{
For more information on the use of the set of functions within the package see the help vignette: \code{vignette("Introduction", package = "corrfx")}
}
