% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initfile.R
\name{initfile}
\alias{initfile}
\title{init file: function used to create subsets of larger dataset and write files for each species for raw, log, and rank transformations}
\usage{
initfile(data, grpCol, dataType, returnCSV = T, pathFolder)
}
\arguments{
\item{data}{a dataframe to be split. This dataframe should have the species code as "spcode"}

\item{grpCol}{A string. name of column with species codes in it.}

\item{dataType}{A string. Indicate what data type label for output csv files}

\item{returnCSV}{TRUE or FALSE. Do you wish to have subsets output as csv files? Default is TRUE}

\item{pathFolder}{A string passed to here function indicate folder for file placement}
}
\value{
Returns a list of group specific subsets of the larger data frame and .csv files(optional)
}
\description{
init file: function used to create subsets of larger dataset and write files for each species for raw, log, and rank transformations
}
