% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_corr.R
\name{p_corr_mat}
\alias{p_corr_mat}
\title{Partial correlation extension of the correlation matrix function \code{corr_mat}}
\usage{
p_corr_mat(
  dat = dat,
  k,
  raw = F,
  rank = F,
  log = F,
  file1 = "correlation matrix file name",
  file2 = "p value file name"
)
}
\arguments{
\item{dat}{a data frame}

\item{k}{variable column index to be \code{partialed} out or accounted for}

\item{raw}{True or False; is your input data frame \code{dat} raw}

\item{rank}{True or False; is your input data frame \code{dat} log-transformed}

\item{log}{True or False; is your input data frame \code{dat} rank transformed}

\item{file1}{correlation matrix file name}

\item{file2}{p value file name}
}
\value{
Returns a correlation matrix partialing out the selected variable k
}
\description{
Partial correlation extension of the correlation matrix function \code{corr_mat}
}
